+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Release NEXT:
  - `OidcSiopWallet`: Rename `newInstance()` to `newDefaultInstance()`, to align it with other factory methods
  - `OidcSiopWallet`: Rename `retrieveAuthenticationRequestParameters() ` to `parseAuthenticationRequestParameters()`, changing result type to `KmmResult<AuthenticationRequestParameters>`
  - `OidcSiopWallet`: Support getting presentation definition remotely, with `presentation_definition_uri` from OpenId4VP
+ - Be more lenient when parsing several authentication request parameters
 
 Release 3.6.1:
  * Update to KMP-Crypto 2.6.0